requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
boto3==1.34.100
botocore==1.34.100
certifi==2024.2.2
charset-normalizer==3.3.2
dj-database-url==2.1.0
Django==5.0.1
django-ckeditor-5==0.2.13
django-cleanup==8.1.0
django-csp==4.0b1
django-js-asset==2.2.0
django-stdimage==6.0.2
django-storages==1.14.3
environs==11.0.0
filelock==3.13.1
gunicorn==22.0.0
idna==3.6
jmespath==1.0.1
marshmallow==3.21.1
packaging==24.0
pillow==10.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
requests-file==2.0.0
s3transfer==0.10.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tldextract==5.1.2
typing_extensions==4.11.0
tzdata==2023.4
urllib3==2.2.1
whitenoise==6.6.0
